[Function][Integrate] Refactoring how binding editor gets opened/closed
