+++ client-react/src/pages/app/functions/integrate/BindingsDiagram/BindingDiagram.styles.ts
