+++ client-react/src/pages/app/functions/integrate/BindingsDiagram/OutputBindingCard.tsx
@@ -1,17 +1,17 @@
 import React from 'react';
-import DataFlowCard, { DataFlowCardChildProps, getBindings } from './DataFlowCard';
+import BindingCard, { BindingCardChildProps, getBindings } from './BindingCard';
 import { ReactComponent as OutputSvg } from '../../../../../images/Common/output.svg';
 import { useTranslation } from 'react-i18next';
 import { BindingDirection } from '../../../../../models/functions/function-binding';
 
-const OutputDataFlowCard: React.SFC<DataFlowCardChildProps> = props => {
+const OutputBindingCard: React.SFC<BindingCardChildProps> = props => {
   const { functionInfo } = props;
   const { t } = useTranslation();
 
   const outputs = getBindings(functionInfo.properties.config.bindings, BindingDirection.out);
 
   return (
-    <DataFlowCard
+    <BindingCard
       items={outputs}
       title={t('output')}
       emptyMessage={t('integrateNoOutputsDefined')}
@@ -22,4 +22,4 @@ const OutputDataFlowCard: React.SFC<DataFlowCardChildProps> = props => {
   );
 };
 
-export default OutputDataFlowCard;
+export default OutputBindingCard;
